% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jags-Ymet-Xnom1fac-MnormalHom.R
\name{genMCMC_ANOVA}
\alias{genMCMC_ANOVA}
\title{Bayesian ANOVA with homogenous variance}
\usage{
genMCMC_ANOVA(datFrm, yName = "y", xName = "x", numSavedSteps = 50000,
  thinSteps = 1, saveName = NULL)
}
\description{
Bayesian ANOVA with homogenous variance
}

