% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jags-Ycount-Xnom2fac-MpoissonExp.R
\name{genMCMC_cont_table}
\alias{genMCMC_cont_table}
\title{Bayesian contingency table test}
\usage{
genMCMC_cont_table(datFrm, yName = "y", x1Name = "x1", x2Name = "x2",
  numSavedSteps = 50000, thinSteps = 1, saveName = NULL,
  showSDprior = FALSE)
}
\description{
Bayesian contingency table test
}

